Improve printing of hoogle output.
Currently, when hoogle completes, the buffer is scrolled so that
'[Process hoogle finished]' is at the top, and looks like no results
are returned. Scrolling back up fixes this, but it's nicer to just
return point to the top of the buffer after hoogle completes. This
patch does just that.