Add texinfo manual skeleton
This is mostly to set up the infrastructure needed for distributing an
Info manual. This commit also adds a few references from docstrings to
Info nodes for testing.

When installing via MELPA/Marmalade, the info file location should be
automatically registered, however when working directly from git
checkout the following is needed:

    (add-to-list 'Info-default-directory-list "~/<some-folder...>/haskell-mode/")