Add new `haskell-str` module
This module is going to replace `haskell-string` in the long run; it
features a slightly shorter namespace function prefix `haskell-str-*`
and tries to better follow Elisp coding conventions.

Most notably, this module provide functions for converting to/from
Haskell string literals, which sets up some of the infrastructure
needed for addressing #164 and #178.