New HASKELL-PROCESS-INSERT-TYPE function for HASKELL-PROCESS-DO-TYPE.
The problem with the new presentation mode for C-c C-t was that C-u
would cause HASKELL-PROCESS-DO-SIMPLE-ECHO to try to insert the type
into the read-only buffer.

Given that you probably don't want to echo when you're inserting a
type with C-u C-c C-t, I separated out the insertion code from the
simple query code, into HASKELL-PROCESS-INSERT-TYPE, which also stores
the current buffer so that it will insert into the right buffer (this
is useful for keyboard macros).

I also added some detection in there to not bother inserting
errors.

A few things can be abstracted into functions, will do in future
commits.

@hvr