Don't jump a whole line on DEL at the leftmost indentation
The previous behavior was to jump to the previous line by hitting DEL
just once when there were no indentation points to the left anymore.
This was quite confusing and almost never intended from the user's
point of view.  If the user is still backspacing when there are no
indentation points left, chances are that we screwed up something
already and the user is intending to indent somewhere unknown.  Better
to just fallback to regular DEL behavior then.