Refactor `haskell-cabal-{compute-checksum,find-file}'
This commit adds proper docstrings and refactors the two functions in
terms of the new `haskell-cabal-find-pkg-desc' function added in
5b02b04. Because by using `haskell-cabal-find-pkg-desc' the
implementation becomes more faithful to the way Cabal locates `.cabal`
files, there are subtle semantic changes caused by this
commit. However, since the goal is to emulate Cabal as close as
possible, this should be alright.