more precise .gitignore
Be more precise about what should be ignored.

Note that *~ belongs in per-user configuration, as it depends on the
editor being used.  It might seem unlikely that vi users will be
working on this particular package, but it's the principle of the
thing!  The right way to do this is:

$ (echo "*~"; echo ".#*[0-9]") > ~/.gitignore-emacs
$ git config --global core.excludesfile ~/.gitignore-emacs