Automatically add import lines via Hoogle
In similar vein to the other suggestion functionality, now, when "Not
in scope: `foo'" is seen, it will search hoogle for "foo" and if
hoogle returns any result "Foo.Bar foo :: a -> b" it will prompt:

Identifier `foo' not in scope, import `Foo.bar'? (y or n)

And then automatically insert an import for that module in the
offending file. The Hoogle executable needs to be installed for this,
which can be installed via

$ cabal install hoogle

and its database needs to be populated with

$ hoogle data

which will download some standard indexes from the Hackage servers. If
you want Hoogle on all your installed packages and your local
codebase, it's not straight-forward, for reasons unknown.

@hvr might be of interest.