Colorize hoogle output, and run hoogle synchronously
Hoogle searches locally and runs quickly, so it rapidly produces all
its output. It's better, then, to insert it in one shot into the
result buffer (now called "*hoogle*" instead of "*help*").

And if we pass the "--color" option, we can then get more readable
output, provided we pass it through "ansi-color-apply-on-region".

Closes #299