Fix haskell-process-load-file
Fixes a bug in which if buffer-file-name had a space in it, a "not a
module name or a source file" error would happen. This was caused by
":load" argument not being quotes.

This commit quotes the argument and replaces all "\"" with "\\"" - just
in case you are some crazy person who uses quotes in their file names.