Propertize haskell-interactive-mode with 'font-lock-face, not 'face
Font lock is assumed to be enabled in haskell-interactive-mode, since it
is used to colorize ":t" output etc.  As part of its operational cycle,
font-lock translates the 'font-lock-face property into 'face,
overwriting what might be there already. It is therefore generally
better to propertize text with 'font-lock-face than 'face, to avoid
clashes.

For example, the ansi-color code used in the previous commit uses
'font-lock-face, which is why the accompanying 'face property was
renamed. This commit extends the principle to this entire file.