Add preliminary :loc-at based haskell-mode-goto-loc
@hvr We now have a number of ways to jump to things. I'll sit down and
consolidate these at some point. Technically all of them are different
and have different reliabilities, so doing a staged fallback of:

* :loc-at
* :i
* find-tag

Seems like a pretty robust way to almost always get something
useful. But for the time being before doing that I'll play around with
this and check for obvious bugs.