Put Haskell symbol chars in punctuation char class.
Emacs defines 'w' class as word chars and '_' as set that together with
'w' can form identifiers. Note that by this operators in programming
languages fall under punctuation '.' class and are treated as such by
for example c-mode, java-mode, etc.

It is confusing that characters defined in Haskell Report as 'symbol'
are NOT the same as 'symbol' character class defined in Emacs.