Find Cabal file even when directory does not yet exist
If you have a Cabalised project in an existing directory
"/path/to/project/" and you want to add a new module into an as-yet
nonexistent directory "Foo", you still might want to be able to find the
.cabal file for that project (e.g. to use fields from the .cabal file to
populate a skeleton for the new module).

This patch keeps stripping off directories from the end of the path
until an existing directory is found (or the directory is found to be
nonsensical).