Use faces directly.
Use faces directy as symbols not as variables. As per font-lock
documentation:

Note that in new code, in the vast majority of cases there is no need to
create variables that specify face names.  Simply using faces directly
is enough.  Font-lock is not a template to be followed in this area.