Made `haskell-ident-pos-at-point` return `nil`.
The documentation mandates that `haskell-ident-pos-at-point` return nil
when there is no identifier at point; however, it instead returned a
cons cell of the form `(n . n)`.  This is now fixed, and any code that
relied on this behavior has been changed.

This also therefore fixes `haskell-spanable-pos-at-point` in the same
way.