Added tests for `haskell-ident-at-point` & friends
As gracjan requested in #603, I

> * add[ed] a test case for haskell-ident-at-point really returning nil when it should
> * add[ed] a test case for haskell-ident-pos-at-point returning non-nil
> * add[ed] a test case for haskell-ident-pos-at-point returning nil
> * add[ed] a test case for haskell-spanable-pos-at-point returning non-nil
> * add[ed] a test case for haskell-spanable-pos-at-point returning nil

I added some somewhat detailed `nil`-checking, a couple quick tests for
`haskell-{ident,spanable}-pos-at-point` that just try the same unicode
identifiers, and then some tests to distinguish
`haskell-spanable-pos-at-point` by using an identifier in backticks.

Any test for `haskell-ident-at-point` could also be a test for the other
two, so that's a potential source of more tests; however, the
definitions of the functions in terms of `haskell-ident-pos-at-point`
are so simple that this didn't seem to add much (especially given the
sheer amount of test duplication).