preventing the unused variable warning.
The warning report system of the byte-compiler is not perfect.
It generates the warning if the third argument of dotimes is used.
Also, it does so if the first argument is not explicitly used in its body.
As workaround, this patch removes the third argument and put "_" to
the first argument.