Fix hoogle lookup from local servers
Use a better way to check if the hoogle process is alive: Instead of
checking if the corresponding buffer-name is an alive process, store the
process handle in a global variable and check that.

The problem is that when using 'hoogle-lookup-from-local' without any
hoogle server running, a new local hoogle server is started, but the
code fails to recognize that server is running and instead repeatedly
reports that the server is not running offering to start it.