Don't spuriously set mark on xref push
This fixes #863, whereby jumping to the definition of a symbol in the
same file was incorrectly activating the mark and region. This commit
also simplifies the original code, since the initial position marker can
simply be passed to xref-push-marker-stack. (The haskell-compat shim for
that function is also fixed here accordingly.)