Detect and respect the comma style of a section
Previously, `haskell-cabal-list-comma-position' was always used when
calling `haskell-cabal-subsection-arrange-lines'.

In #1098, `haskell-cabal-list-comma-position' was respected unless the
no commas were found.

Instead of respecting `haskell-cabal-list-comma-position', detect the
currently used style and respect it when rearranging the section.

Possible styles:

before: a comma at the start of each line (except the first), e.g.
    Foo
  , Bar

after: a comma at the end of each line (except the last), e.g.
    Foo,
    Bar

single: everything on a single line, but comma-separated, e.g.
    Foo, Bar

nil: no commas, e.g.
    Foo Bar

If the styles are mixed, the position of the first comma determines the
style.