use `haskell-interactive-switch` in process-load
This sets a variable with the source buffer so we can switch back.

This is pretty bizzare, though.  Both before and after this commit,
`haskell-session-interactive-buffer` switches to the buffer if it's
newly created, but if the buffer already exists, just returns the buffer
without visiting.  This function shouldn't switch, and callers
should decide if they want to switch.

closes #955