Fix #1146 by treating . as punctuation
In `haskell-cabal-mode` with the point indicated with the vertical bar
below, `forward-char` should leave the point after "Some" instead of
after "Type".

    Exposed-Modules:   |Some.Internal.Type

This was caused by the following line:

    (modify-syntax-entry ?. "w"  st)

Which makes "." a word constituent instead of a punctuation character.
This was added in aa970092a1e812ff78af2c1d65484b4281ce2234, "merge
caballero with haskell-cabal". The fix: remove the line above.