Rename some functions
1. Renamed `haskell-cabal-get-setting` to `haskell-cabal--get-field`.
Word "field" was chosen because this word is used in cabal's user guide.
Double dash was used because this function is likely designed for
internal usage according to its logics, e.g. it do makes no checks and
just blindly searches for `:NAME` occurance in current buffer

2. `haskell-cabal-guess-setting` was renamed to
`haskell-cabal-get-field` because it does not guesses anything, but
simply tries to find cabal file and if succeeds calling
`haskell-cabal--get-field`.