Demonstrate the problem with haskell-process-do-cabal.
haskell-process-do-cabal has code that attempts to run cabal directly
when there is no current interactive haskell process.  Unfortunately, it
doesn't work.

This test demonstrates the problem, calling haskell-process-do-cabal and
checking the *Messages* buffer for the presence of the string it should
output if it was going to run cabal directly.

Instead you get an error about the lack of a session associated with the
buffer.