Handle cabal indentation relative to first line
Previously, cabal-mode indented to the ":" column of a subsection in a
cabal file. This was not how most users indented their .cabal files
and was surprising. Instead of doing this, we indent relatively to the
first entry in a subsection list. When it doesn’t exist, use the old
behavior.

:data-indent-column is added to store this new indent level. It needs
to be separate from :data-start-column due to :data-start-colum also
being used in finding the start of a subsection heading.