Replaces customize var. haskell-compile-ignore-cabal by haskell-compiler-type,
which is similar to haskell-process-type. This makes haskell-compile
choose build tools the same way haskell-interactive-mode does.

Values are 'cabal, 'cabal-project, 'stack, or 'ghc to choose a build tool,
or 'auto (default) to look for stack or cabal files.

For haskell-process-type the corresponding values remain 'cabal-repl,
'cabal-new-repl, 'stack-repl, 'ghci.

In fact 'cabal-new-repl is obsolete and is accepted as a setting of the
defcustom, with a warning, but the accessor function (haskell-process-type)
returns 'cabal-repl instead.

Factors out logic that deals with value 'auto for both of these, as
(haskell-build-type). Also removes a bit of duplicated code from
(haskell-process-do-cabal) in haskell-load.el.