Reduce (over)use of markers (#1751)
The parser used by the font-lock code created a massive amount of
markers; try to reduce it as far as possible since they make some
operations expensive as long as they are in the buffer. Even orphaned
markers stay until next GC.

Read match data as integers instead of markers, and don't create new
markers when setting the match data. This should work just fine as
long as the buffer isn't changed during the match data lifetime, which
appears reasonable.