Add interactive command to insert a LANGUAGE pragma
I need to manually add `{-# LANUGAGE _ #-}` pragmas to the top of my modules pretty often, so I wrote a function to do this. I've been using it for a few days and found it really convenient, and other people liked the idea when I mentioned it on Twitter.

Called interactively it uses `haskell-ghc-supported-extensions` for completions which means less typing and no more errors like using `RecordWildcards` instead of `RecordWildCards`.

I bound it to `C-c C-e` on my own setup, but not sure whether it makes sense to include a keybinding for it by default.