haskell-cabal: fix (comment-dwim) by correcting comment-start-skip
Prior to this commit using (comment-dwim) over a comment in .cabal
file that doesn't start at the beginning of a line resulted in it
adding even more comments. E.g. this:

    -- comment not at the beginning of a line

resulted in:

    -- -- comment not at the beginning of a line

This is because `comment-start-skip` was including whitespace that
isn't part of the comment syntax. Fix this by only limiting the regexp
to the comment part.

The resulting regexp is basically same as the one in haskell-mode,
barring that that one also takes into account `{-` comment starter,
which isn't a thing in .cabal files.

Fixes: https://github.com/haskell/haskell-mode/issues/1743